requirements.txt
# Required Python packages for SCD Project
psycopg2-binary==2.9.9==2.9.9
